+++ package.json
@@ -19,7 +19,7 @@
   "dependencies": {
     "@babel/runtime": "^7.12.5",
     "@bit/joshk.vue-spinners-css.spinner-loader": "^1.1.5",
-    "@debionetwork/pinata-ipfs": "0.0.1",
+    "@debionetwork/pinata-ipfs": "0.0.2",
     "@debionetwork/polkadot-provider": "0.0.6",
     "@debionetwork/ui-components": "0.0.5",
     "@debionetwork/ui-icons": "0.0.2",
