+++ src/views/Dashboard/Customer/Home/Emr/Details.vue
@@ -48,7 +48,7 @@ import Kilt from "@kiltprotocol/sdk-js"
 import CryptoJS from "crypto-js"
 import { u8aToHex } from "@polkadot/util"
 import { fileTextIcon } from "@debionetwork/ui-icons"
-import { downloadFile, decryptFile } from "@/common/lib/pinata"
+import { downloadFile, decryptFile } from "@/common/lib/pinata-proxy"
 import {
   queryElectronicMedicalRecordById,
   queryElectronicMedicalRecordFileById
