+++ src/views/Dashboard/Customer/Home/GeneticData/GeneticAnalysisList.vue
@@ -35,7 +35,7 @@ import { queryGeneticAnalystByAccountId } from "@debionetwork/polkadot-provider"
 import { queryGeneticAnalysisByOwnerId } from "@debionetwork/polkadot-provider"
 import { queryGeneticAnalysisStorage, queryGeneticAnalysisOrders } from "@/common/lib/polkadot-provider/query/genetic-analysis"
 import { queryGeneticAnalystServices } from "@/common/lib/polkadot-provider/query/genetic-analysts"
-import { downloadFile, decryptFile, downloadDocumentFile } from "@/common/lib/pinata"
+import { downloadFile, decryptFile, downloadDocumentFile } from "@/common/lib/pinata-proxy"
 import Kilt from "@kiltprotocol/sdk-js"
 import { u8aToHex } from "@polkadot/util"
 import CryptoJS from "crypto-js"
