+++ src/views/Dashboard/Customer/Home/GeneticData/RequestAnalysis/AnalystDetail.vue
@@ -88,7 +88,7 @@ import Kilt from "@kiltprotocol/sdk-js"
 import CryptoJS from "crypto-js"
 import cryptWorker from "@/common/lib/ipfs/crypt-worker"
 import { u8aToHex } from "@polkadot/util"
-import { downloadFile, uploadFile, getFileUrl } from "@/common/lib/pinata"
+import { downloadFile, uploadFile, getFileUrl } from "@/common/lib/pinata-proxy"
 import { createGeneticAnalysisOrder } from "@debionetwork/polkadot-provider"
 import { queryGeneticAnalysisOrderByCustomerId } from "@debionetwork/polkadot-provider"
 import { queryGeneticAnalystByAccountId } from "@debionetwork/polkadot-provider"
